✂️ Code refactoring: Splitting the main code into separate files
In this commit, the main code of the application has been refactored and split into separate files for better organization and maintainability. The main function and the main class of the application are now in the main.dart file, the state management class is in the app_state.dart file, the home page class is in the my_home_page.dart file, the generator page class is in the generator_page.dart file, and the favorites page class is in the favorites_page.dart file. This change allows for easier debugging, code updates and better collaboration between team members.
