+++ lib/app_state.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+import 'package:english_words/english_words.dart';
+
+class MyAppState extends ChangeNotifier {
+  var current = WordPair.random();
+  var history = <WordPair>[];
+
+  GlobalKey? historyListKey;
+
+  void getNext() {
+    history.insert(0, current);
+    var animatedList = historyListKey?.currentState as AnimatedListState?;
+    animatedList?.insertItem(0);
+    current = WordPair.random();
+    notifyListeners();
+  }
+
+  var favorites = <WordPair>{};
+
+  void toggleFavorite([WordPair? pair]) {
+    pair = pair ?? current;
+    if (favorites.contains(pair)) {
+      favorites.remove(pair);
+    } else {
+      favorites.add(pair);
+    }
+    notifyListeners();
+  }
+
+  void removeFavorite(WordPair pair) {
+    favorites.remove(pair);
+    notifyListeners();
+  }
+}
